refactor(components): refactor page-header
